BeautifulNumbers problem added

+++ src/algorithms/algebra/fundamentals/binaryExponentiation/BinaryExponentiation.kt
@@ -8,12 +8,12 @@ package algorithms.algebra.fundamentals
  */
 fun binExpRec(a: Long, n: Long): Long {
     tailrec fun go(a: Long, n: Long, acc: Long): Long{
-        if (a == 0.toLong()) return 0
-        if (n == 0.toLong()) return 1
-        if (n == 1.toLong()){
+        if (a == 0L) return 0
+        if (n == 0L) return 1
+        if (n == 1L){
             return a * acc
         }
-        if (n % 2 == 0.toLong())
+        if (n % 2 == 0L)
             return go(a * a, n/2, acc)
         else
             return go(a * a, (n - 1)/2, a * acc)
@@ -23,7 +23,7 @@ fun binExpRec(a: Long, n: Long): Long {
 
 fun binExpIter(a: Long, n: Long): Long{
 
-    if (a == 0.toLong()) return 1
+    if (a == 0L) return 1
 
     var x = a
     var y: Long = 1
@@ -31,7 +31,7 @@ fun binExpIter(a: Long, n: Long): Long{
 
     while (count > 1){
 
-        if (count % 2 == 0.toLong()){
+        if (count % 2 == 0L){
             count /= 2
         }else{
             y *= x
@@ -50,14 +50,14 @@ fun binExpIter(a: Long, n: Long): Long{
  * Reminder: a * b = ((a mod m) * (b mod m)) mod m
  */
 fun binExpMod(a: Long, n: Long, m: Long): Long{
-    if (a == 0.toLong()) return 1
+    if (a == 0L) return 1
 
     var x = a
     var y: Long = 1
     var count = n
 
     while (count > 1){
-        if (count % 2 == 0.toLong()) count /= 2
+        if (count % 2 == 0L) count /= 2
         else{
             y = (y * x) % m
             count = (count - 1)/2
@@ -70,11 +70,11 @@ fun binExpMod(a: Long, n: Long, m: Long): Long{
 
 fun binExpMod1(a: Long, n: Long, m: Long): Long{
     tailrec fun go(a: Long, n: Long, acc: Long): Long{
-        if (a == 0.toLong()) return 0
-        if (n == 1.toLong()){
+        if (a == 0L) return 0
+        if (n == 1L){
             return (a * acc) % m
         }
-        if (n % 2 == 0.toLong())
+        if (n % 2 == 0L)
             return go((a * a) % m, n/2, acc)
         else
             return go((a * a) % m, (n - 1)/2, (a * acc) % m)
@@ -91,11 +91,11 @@ fun binExpMod1(a: Long, n: Long, m: Long): Long{
  */
 fun multRec(a: Long, b: Long): Long{
     tailrec fun go(a: Long, b: Long, acc: Long): Long{
-        if (a == 0.toLong() || b == 0.toLong()) return 0
-        if (a == 1.toLong()){
+        if (a == 0L || b == 0L) return 0
+        if (a == 1L){
             return b + acc
         }
-        if (a % 2 == 0.toLong())
+        if (a % 2 == 0L)
             return go(a/2, b * 2, acc)
         else
             return go((a - 1)/2, b * 2, b + acc)
@@ -108,11 +108,11 @@ fun multRec(a: Long, b: Long): Long{
  */
 fun multMod(a: Long, b: Long, m: Long): Long{
     tailrec fun go(a: Long, b: Long, acc: Long): Long{
-        if (a == 0.toLong() || b == 0.toLong()) return 0
-        if (a == 1.toLong()){
+        if (a == 0L || b == 0L) return 0
+        if (a == 1L){
             return (b + acc) % m
         }
-        if (a % 2 == 0.toLong())
+        if (a % 2 == 0L)
             return go(a/2, (b * 2) % m, acc)
         else
             return go((a - 1)/2, (b * 2) % m, (b + acc) % m)
