+++ src/algorithms/algebra/fundamentals/greatestCommonDivisor/GreatestCommonDivisor.kt
@@ -2,7 +2,7 @@ package algorithms.algebra.fundamentals.greatestCommonDivisor
 
 //data class Tuple4<A, B, C, D>(val first: A, val second: B, val third: C, val fourth: D)
 
-val zero = 0.toLong()
+val zero = 0L
 
 /**
  * g = gcd(a, b) => (g | gcd(b, a mod b))  <Divides>
